HMGit.Internal.Core

Documentation

type HMGitT = ReaderT HMGitConfig

hmGitRoot :: Monad m => HMGitT m (Path Abs Dir)

getCurrentDirFromHMGit :: (MonadThrow m, MonadIO m) => HMGitT m (Path Rel Dir)

runHMGit :: HMGitT m a -> HMGitConfig -> m a

data ObjectType

Constructors

Blob

Commit

Tree

data ObjectInfo

Constructors

ObjectInfo

Fields

objectId :: ByteString

objectData :: ByteString

objectPath :: Path Abs File

data IndexEntry

Constructors

IndexEntry

Fields

ieCtimeS :: Word32

the last time a file's metadata changed, this is stat(2) data

ieCtimeN :: Word32

nanosecond fractions, this is stat(2) data

ieMTimeS :: Word32

mtime seconds, the last time a file's data changed, this is stat(2) data

ieMTimeN :: Word32

mtime nanosecond fractions, this is stat(2) data

ieDev :: Word32

this is stat(2) data

ieIno :: Word32

this is stat(2) data

ieMode :: Word32

mode, split into (high to low bits)

ieUid :: Word32

this is stat(2) data

ieGid :: Word32

this is stat(2) data

ieSize :: Word32

This is the on-disk size from stat(2), truncated to 32-bit.

ieSha1 :: ByteString

160-bit SHA-1 for the represented object

ieFlags :: Word16

A 16-bit flags field split into (high to low bits)

iePath :: Path Rel File

fromContents :: MonadCatch m => ObjectType -> ByteString -> HMGitT m ObjectInfo

storeObject :: (MonadIO m, MonadCatch m) => ObjectType -> ByteString -> HMGitT m ByteString

loadObject :: (MonadIO m, MonadCatch m, MonadPlus m) => String -> HMGitT m (ObjectType, ByteString)

loadTree :: MonadThrow m => ByteString -> HMGitT m [(CMode, Path Rel File, String)]

storeTree :: (MonadIO m, MonadCatch m) => HMGitT m ByteString

loadIndex :: (MonadIO m, MonadThrow m) => HMGitT m [IndexEntry]

storeIndex :: (MonadIO m, Foldable t) => t IndexEntry -> HMGitT m ()

data HMGitStatus

Currently latestBlobHashes does not support gitignore and submodule, so we are embedding content to ignore directly in the code. Comments HACK below are the relevant part.

Constructors

HMGitStatus

Fields

statusChanged :: Set (Path Rel File)

statusNew :: Set (Path Rel File)

statusDeleted :: Set (Path Rel File)

indexedBlobHashes :: (MonadIO m, MonadCatch m) => HMGitT m (Map (Path Rel File) String)

getStatus :: (MonadIO m, MonadCatch m) => HMGitT m HMGitStatus