HMGit.Internal.Parser.Pathspecs

Description

Matching patterns against file paths akin to the pathspec gitglossary(7) function.

NOTE: Currently does not support magic signatures

See also: https://git-scm.com/docs/gitglossary#Documentation/gitglossary.txt-aiddefpathspecapathspec

Documentation

pathspecs

Arguments

:: (MonadCatch m, MonadIO m, Alternative m)

=> Path Abs Dir
The specified base directory

-> SomeBase File
The target file path

-> [String]
pathspecs

-> HMGitT m FilePath

Determines if the path given with cDir as the base point matches the pattern, and if so, returns a FilePath of relative paths (including ../ etc.) from cDir (P.Path is not supported ../, so we must return as a string). If they do not match, some exception will be thrown.

lsMatches

Arguments

:: (MonadCatch m, MonadIO m)

=> Path Abs Dir
the specified base directory

-> [String]
pathspecs

-> HMGitT m (Set (Path Abs File))