HMGit.Internal.Core

Documentation

data ObjectType

The Git object type. Currently, it does not support tags.

Constructors

Blob
Blob object

Commit
Commit object

Tree
Tree object

data ObjectInfo

Information for built objects

data IndexEntry

Constructors

IndexEntry

Fields

ieCtimeS :: Word32

the last time a file's metadata changed, this is stat(2) data

ieCtimeN :: Word32

nanosecond fractions, this is stat(2) data

ieMTimeS :: Word32

mtime seconds, the last time a file's data changed, this is stat(2) data

ieMTimeN :: Word32

mtime nanosecond fractions, this is stat(2) data

ieDev :: Word32

this is stat(2) data

ieIno :: Word32

this is stat(2) data

ieMode :: Word32

mode, split into (high to low bits)

ieUid :: Word32

this is stat(2) data

ieGid :: Word32

this is stat(2) data

ieSize :: Word32

This is the on-disk size from stat(2), truncated to 32-bit.

ieSha1 :: ByteString

160-bit SHA-1 for the represented object

ieFlags :: Word16

A 16-bit flags field split into (high to low bits)

iePath :: Path Rel File

fromContents :: MonadCatch m => ObjectType -> ByteString -> HMGitT m ObjectInfo

Create object ID, path, zlib compressed data from object type and content

storeObject :: (MonadIO m, MonadCatch m) => ObjectType -> ByteString -> HMGitT m ByteString

Create an object to HMGit database

loadObject :: (MonadIO m, MonadCatch m, MonadPlus m) => String -> HMGitT m (ObjectType, ByteString)

Read an object from HMGit database by Object ID

loadTree :: MonadThrow m => ByteString -> HMGitT m [(CMode, Path Rel File, String)]

Read a tree from raw data

storeTree :: (MonadIO m, MonadCatch m) => HMGitT m ByteString

Create a tree object to HMGit database NOTE: Currently only supports a single, top-level directory

loadIndex :: (MonadIO m, MonadThrow m) => HMGitT m [IndexEntry]

Read index entries

storeIndex :: (MonadIO m, Foldable t) => t IndexEntry -> HMGitT m ()

Write index entries

data HMGitStatus

Data representing status

latestBlobHashes :: (MonadIO m, MonadCatch m) => HMGitT m (Map (Path Rel File) String)

Calculates the latest, or current, hash of the working tree and returns a map of its path and hash value.

NOTE: Currently latestBlobHashes does not support gitignore and submodule, so we are embedding content to ignore directly in the code. Comments HACK below are the relevant part.

indexedBlobHashes :: (MonadIO m, MonadCatch m) => HMGitT m (Map (Path Rel File) String)

Returns a map of blob file paths and hashes that are already part of the repository. This is a pair with latestBlobHashes.

getStatus :: (MonadIO m, MonadCatch m) => HMGitT m HMGitStatus

Get status